+++ src/components/Explorer/index.tsx
@@ -19,7 +19,8 @@ import Form from '../Form';
 import getValidation from '../Form/schema';
 import skip from '../lib/skip';
 import fieldNames from '../lib/fields';
-import columnProps, {TableFilter} from '../lib/column';
+import columnProps from '../lib/column';
+import useFilter from '../hooks/useFilter';
 import prepareSubmit from '../lib/prepareSubmit';
 
 import { ComponentProps } from './Explorer.types';
@@ -44,6 +45,10 @@ const useStyles = createUseStyles({
             width: 'fit-content'
         },
         '& .p-datatable-tbody .p-button .p-button-label': {
+            overflow: 'hidden',
+            textOverflow: 'ellipsis'
+        },
+        '& .p-datatable-tbody td .value': {
             overflow: 'hidden',
             textOverflow: 'ellipsis'
         },
@@ -183,7 +188,7 @@ const Explorer: ComponentProps = ({
     if (toolbar !== false) toolbar = ('layout' in layoutProps) ? ('toolbar' in layoutProps ? mergedCards[layoutProps.toolbar]?.widgets : toolbar) : mergedCards[toolbarCard]?.widgets ?? toolbar;
     const classes = useStyles();
     const {properties} = mergedSchema;
-    const [tableFilter, setFilters] = React.useState<TableFilter>({
+    const [tableFilter, setFilters, filterBy, filterProps] = useFilter({
         filters: columns?.reduce((prev : object, column) => {
             let field = fieldName(column);
             const value = lodashGet(externalFilter, field);
@@ -192,9 +197,8 @@ const Explorer: ComponentProps = ({
         }, {}),
         first: 0,
         page: 1
-    });
+    }, columns, properties, showFilter);
     const multiSelect = keyField && (!tableProps?.selectionMode || tableProps?.selectionMode === 'checkbox');
-    const handleFilterPageSort = React.useCallback(event => setFilters(prev => ({...prev, ...event})), []);
 
     const [{current: currentState, selected: selectedState}, setCurrentSelected] = React.useState({current: null, selected: null});
     const handleCurrentSelect = React.useCallback((value, event) => {
@@ -265,7 +269,7 @@ const Explorer: ComponentProps = ({
             setLoading('');
         }
         if (system?.fetch) setFilters(prev => merge({}, prev, system.fetch));
-    }, [methods, getValues]);
+    }, [methods, getValues, setFilters]);
 
     const handleSubmit = React.useCallback(params => {
         if (params?.[2]?.method) {
@@ -415,30 +419,6 @@ const Explorer: ComponentProps = ({
             }</div>
         </SplitterPanel>, [getValues, result, details, detailsOpened, height]);
 
-    const filterBy = (name: string, key: string) => e => {
-        const value = lodashGet(e, key);
-        setFilters(prev => {
-            const next = {
-                ...prev,
-                filters: {
-                    ...prev?.filters,
-                    [name]: {
-                        ...prev?.filters?.[name],
-                        value: value === '' ? undefined : value
-                    }
-                }
-            };
-            return next;
-        });
-    };
-
-    const filterDisplay = React.useMemo(() => showFilter && columns.some(column => {
-        const isString = typeof column === 'string';
-        const {name, ...widget} = isString ? {name: column} : column;
-        const property = lodashGet(properties, name?.replace(/\./g, '.properties.'));
-        return !!property?.filter || widget?.column?.filter;
-    }), [columns, properties, showFilter]) ? 'row' : undefined;
-
     const [Columns, errorsWithoutColumn] = React.useMemo(() => {
         const errorsWithoutColumn = filterErrors ? [...filterErrors.details] : [];
         return [columns.map((column, index) => {
@@ -450,25 +430,6 @@ const Explorer: ComponentProps = ({
             return (
                 <Column
                     key={name}
-                    body={action && (row => <ActionButton
-                        {...testid(`${resultSet || 'filterBy'}.${field}Item/${row && row[keyField]}`)}
-                        label={row[field]}
-                        className='p-button-link p-0'
-                        action={action}
-                        submit={submit}
-                        params={widget.params ?? property?.params}
-                        getValues={() => ({
-                            filter: externalFilter,
-                            id: row && row[keyField],
-                            current: row,
-                            selected: [row]
-                        })}
-                        // onClick={() => property.action({
-                        //     id: row && row[keyField],
-                        //     current: row,
-                        //     selected: [row]
-                        // })}
-                    />)}
                     filter={showFilter && !!property?.filter}
                     sortable={!!property?.sort}
                     {...columnProps({
@@ -484,6 +445,22 @@ const Explorer: ComponentProps = ({
                         errorsWithoutColumn,
                         ...formProps
                     })}
+                    {...action && {
+                        body: row => <ActionButton
+                            {...testid(`${resultSet || 'filterBy'}.${field}Item/${row && row[keyField]}`)}
+                            label={row[field]}
+                            className='p-button-link p-0'
+                            action={action}
+                            submit={submit}
+                            params={widget.params ?? property?.params}
+                            getValues={() => ({
+                                filter: externalFilter,
+                                id: row && row[keyField],
+                                current: row,
+                                selected: [row]
+                            })}
+                        />
+                    }}
                 />
             );
         }), errorsWithoutColumn.filter(Boolean)];
@@ -499,6 +476,7 @@ const Explorer: ComponentProps = ({
         formProps,
         externalFilter,
         submit,
+        filterBy,
         filterErrors
     ]);
     const hasChildren = !!children;
@@ -573,7 +551,7 @@ const Explorer: ComponentProps = ({
                 sortField={tableFilter.sortField}
                 sortOrder={tableFilter.sortOrder}
                 value={items}
-                onPage={handleFilterPageSort}
+                onPage={filterProps.onPage as () => void}
                 itemTemplate={itemTemplate}
                 {...viewProps}
             /> : <DataTable
@@ -584,19 +562,12 @@ const Explorer: ComponentProps = ({
                 rows={pageSize}
                 totalRecords={totalRecords}
                 paginator
-                first={tableFilter.first}
-                sortField={tableFilter.sortField}
-                sortOrder={tableFilter.sortOrder}
-                filters={tableFilter.filters}
-                onPage={handleFilterPageSort}
-                onSort={handleFilterPageSort}
-                onFilter={handleFilterPageSort}
+                {...filterProps}
                 loading={!!loading}
                 dataKey={keyField}
                 value={items}
                 rowClassName={rowClass}
                 selection={selected}
-                filterDisplay={filterDisplay}
                 onSelectionChange={handleSelectionChange}
                 onRowSelect={handleRowSelect}
                 onRowUnselect={handleRowUnselect}